BingSearch/requirements.txt
azure-cognitiveservices-search-websearch